refactor(app): separate tools and extensions settings
