+++ packages/app/src/settings/extensions/extensions.css
